Break apart the swap scene styles
